Satisfy code climate and improve readability.

+++ modules/metastore/src/EventSubscriber/MetastoreSubscriber.php
@@ -65,44 +65,37 @@ class MetastoreSubscriber implements EventSubscriberInterface {
   /**
    * React to a distribution being orphaned.
    *
+   * Removes resources associated with the orphaned distribution.
+   *
    * @param \Drupal\common\Events\Event $event
    *   The event object containing the resource uuid.
    */
   public function cleanResourceMapperTable(Event $event) {
-    $uuid = $event->getData();
+    $distribution_id = $event->getData();
     // Use the metastore service to build a distribution object.
-    $distribution = $this->service->get('distribution', $uuid);
-    $distribution = json_decode($distribution);
-    // Attempt to extract the resource for the given distribution.
-    $resource = $distribution->data->{'%Ref:downloadURL'} ?? [];
-    $resource = array_shift($resource);
-    // Retrieve the distributions ID, perspective, and version metadata.
-    $id = $resource->data->identifier ?? NULL;
-    $perspective = $resource->data->perspective ?? NULL;
-    $version = $resource->data->version ?? NULL;
+    $distribution = $this->service->get('distribution', $distribution_id);
+    // Attempt to extract all resources for the given distribution.
+    $resources = $distribution->{'$.data["%Ref:downloadURL"]..data'} ?? [];
 
-    // Use the metastore resourceMapper to remove the source entry.
-    try {
+    // Remove all resource entries associated with this distribution from the
+    // metadata resource mapper.
+    foreach ($resources as $resource) {
+      // Retrieve the distributions ID, perspective, and version metadata.
+      $resource_id = $resource['identifier'] ?? NULL;
+      $perspective = $resource['perspective'] ?? NULL;
+      $version = $resource['version'] ?? NULL;
       // Ensure a valid ID, perspective, and version were found for the given
       // distribution.
-      if (!isset($id, $perspective, $version)) {
-        throw new \UnexpectedValueException('Distribution does not have resource.');
-      }
-      $resource = $this->resourceMapper->get($id, $perspective, $version);
-      if ($resource) {
+      if (isset($resource_id) && $resource = $this->resourceMapper->get($resource_id, $perspective, $version)) {
+        // Remove resource entry for metadata resource mapper.
         $this->resourceMapper->remove($resource);
       }
       else {
-        $this->loggerFactory->get('metastore')->error('Missing resource');
+        $this->loggerFactory->get('metastore')->error('Failed to remove resource with id "@distribution_id" from source mapping for distribution with id "@resource_id".', [
+          '@distribution_id' => $distribution_id,
+          '@resource_id' => $resource_id,
+        ]);
       }
-    }
-    catch (\Exception $e) {
-      $this->loggerFactory->get('metastore')->error('Failed to remove resource source mapping for @uuid. @message',
-        [
-          '@uuid' => $uuid,
-          '@message' => $e->getMessage(),
-        ]
-      );
     }
   }
